when modifying an embedded element collection within an entity of protected isolation the changes need to be merged into clones correctly.
The code had an execution path where the attribute values were not copied at all, that is when cache policy is protected and the target session is a protected session. These lines of code are identical in functionality with those found in mergeChangesIntoObject of CollectionMapping, DirectCollectionMapping, DirectMapMapping and ObjectReferenceMapping. I assumed when implementing changes the code of AggregateCollectionMapping was forgotten to be updated.

Signed-off-by: Thomas Beckmann (M-Way) <t.beckmann@mwaysolutions.com>